moved additional utils classes
