+++ save-agent/save-agent-common/build.gradle.kts
@@ -39,7 +39,8 @@ kotlin {
 
                 implementation(libs.okio)
                 implementation(libs.ktor.client.core)
-                implementation(libs.ktor.plugins.client.logging)
+                implementation(libs.ktor.client.logging)
+//                implementation(libs.ktor.plugins.client.logging)
                 implementation(libs.kotlinx.coroutines.core)
                 implementation(libs.ktor.client.content.negotiation)
                 implementation(libs.ktor.serialization.kotlinx.json)
