+++ save-agent/save-agent-common/src/commonMain/kotlin/com/saveourtool/save/AgentInitConfig.kt
@@ -1,4 +1,4 @@
-package com.saveourtool.save.agent
+package com.saveourtool.save
 
 import com.saveourtool.save.utils.DEFAULT_SETUP_SH_TIMEOUT_MILLIS
 import kotlinx.serialization.Serializable
