+++ save-agent/save-agent-common/src/commonMain/kotlin/com/saveourtool/save/AgentState.kt
@@ -1,4 +1,4 @@
-package com.saveourtool.save.agent
+package com.saveourtool.save
 
 import kotlinx.serialization.Serializable
